overwrite feature flag selectors to return FFs held in query params
