+++ protocol-designer/src/feature-flags/__tests__/getFlagsFromQueryParams.test.js
@@ -1,5 +1,5 @@
 // @flow
-import { getFlagsFromQueryParams } from '../reducers'
+import { getFlagsFromQueryParams } from '../utils'
 describe('getFlagsFromQueryParams', () => {
   it('should enable the flag passed in via query params when it is set to 1', () => {
     // replace window location search
