+++ protocol-designer/src/feature-flags/index.js
@@ -3,6 +3,7 @@ import { rootReducer, type RootState } from './reducers'
 import * as actions from './actions'
 import * as selectors from './selectors'
 export * from './types'
+export * from './utils'
 
 export { actions, rootReducer, selectors }
